pyproject.toml
[project]
name = "ada"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Donald Thompson", email = "witt3rd@witt3rd.com" }
]
dependencies = [
    "pyperclip>=1.9.0",
    "requests>=2.32.3",
    "markdownify>=0.12.1",
    "pydantic>=2.8.0",
    "elevenlabs>=1.3.1",
    "deepgram-sdk>=3.3.1",
    "python-dotenv>=1.0.1",
    "sounddevice>=0.4.7",
    "openai>=1.35.7",
    "google-generativeai>=0.7.1",
    "vosk>=0.3.44",
    "numpy>=2.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/ada"]
